hv: Enable the compiler warning as error for HV
Fix the compiler warning and turn on the flag to make
compiler warning as compiler error.

Tracked-On: #1343
Acked-by: Anthony Xu <anthony.xu@intel.com>

+++ hypervisor/bsp/uefi/uefi.c
@@ -66,7 +66,7 @@ void *get_rsdp_from_uefi(void)
 	if (!efi_initialized)
 		efi_init();
 
-	return hpa2hva(efi_ctx->rsdp);
+	return hpa2hva((uint64_t)efi_ctx->rsdp);
 }
 
 void *get_ap_trampoline_buf(void)
